+++ src/hooks/useRefreshAlert/index.ts
@@ -1,0 +1,17 @@
+import { useEffect } from "react";
+
+export const useRefreshAlert = () => {
+  useEffect(() => {
+    const unloadCallback = (event: BeforeUnloadEvent) => {
+      event.preventDefault();
+      event.returnValue = "";
+      return "";
+    };
+
+    window.addEventListener("beforeunload", unloadCallback);
+
+    return () => window.removeEventListener("beforeunload", unloadCallback);
+  }, []);
+
+  return null;
+};
