+++ src/pages/citizenship/checkout/confirmed/View/TransactionDetails/index.tsx
