+++ src/pages/citizenship/checkout/index.page.tsx
@@ -1,8 +1,11 @@
 import Head from "next/head";
+import { useRefreshAlert } from "~/hooks/useRefreshAlert";
 import { FactionGuard } from "../FactionGuard";
 import { View } from "./components/View";
 
 const Citizenship = () => {
+  useRefreshAlert();
+
   return (
     <>
       <Head>
